コードレビューを受けての修正

+++ backend/app/models/money_model.py
@@ -3,7 +3,7 @@ from pydantic import BaseModel
 
 class RegisterSalary(BaseModel):
     year_month: str
-    monthly_income: int
+    monthly_income: float
 
 
 class YearMonth(BaseModel):
